Bs -> phi mumu following arXiv 1502.05509
BsToPhiDilepton: Add two public functions to retrieve the integrated and differential angular coefficients.
Adjusted the prefactors for the J_i to use the convention from Virto et al in 1502.05509.
This also means that some pre-factors in functions change slightly.

BsToPhiDileptonAndConjugate: This class gets more changes as it will be the main player for phi mumu, having access to both J, Jbar, h, and s.
Added new properties m_y, m_x, m_gamma_s corresponding to the DeltaGamma/Gamma_s, DeltaM/Gamma_s and Gamma_s.

Added new struct: AngularsCoefficients which mirrors AngularhCoefficients.
Added calculation of the s_i coefficients.

Modified AngularhCoefficients to use the same prefactors as in Virto et al 1502.05509.
Fixed a typo in h_1c.
Added functions for the integrated and differential angular coefficients S, A, K, W, H, and Z.
The K and W correspond to the "S" and "A" in the Kst case and used to be called S and A for the phi case as well.
The new S and A are based on Virto et al 1502.05509 and reflect on the time-integration needed to be performed which introduces factors of y / x and h / s.
The definition of the functions is handled using macros to avoid typos and repetetive typing.

One thing to note which is relevant for the h_i and s_i: the phase exp(i phi_s) is changed to use the same definition as in flavio where e^(i phi) is the phase of -q/p. I am not sure if this correct as Virto et al in 1502.05509 define q / p = e^(i phi_s).