Add all new coefficients to observables.cc
Defined macros to easen the burden of typing. currently not used, todo.

B_s->phi_ll::Gamma now uses the implementation of BsToPhiDileptonAndConjugate reflecting on Virto et al in 1502.05509.
Add all K,W,H,Z,S, and A observables in integrated and differential as well as "theory" and "LHCb" convention.
Change between "theory" and "LHCb" is a minus sign for coefficients 4, 6s, 6c, 7, and 9, see Figure 4 in 1506.03970